Remove filter function, it's available within later versions of guile.
The current implementation of filter using foldr is not tail recursive, and
so can cause stack overflow for very long lists.
It could be re-written to use foldl, however as filter has been available
since guile 1.6 (at least) and the current cgen code uses other function(s)
(for example make-object-property) that are not available in pre-1.6 guile,
I've just remove the definition of filter.

cgen/ChangeLog:

	* utils.scm (filter): Remove as this is available in guile since
	at least version 1.6.