Update format and dates of generated copyright header
The preferred format for the copyright header has changed slightly, and
the dates we generate have lagged behind the actual date.  This commit
updates the format, and the date so that when the opcodes directory of
binutils-gdb is regenerated all of the copyright headers match those
that are actually in the files.

ChangeLog:

	* utils.scm (copyright-fsf): Update format and dates in generated
	copyright header.