Add operand validation support
Consider an register set with the encodings like this:

    (r1 1) (r2 2)

We use two bits to encode these two registers the binary encodings for
the registers are either '01' or '10'.

Next, imagine two instructions like this:

    insn1 REG        00rr
    insn2            0011

where the encoding for 'insn1 r1' would be '0001' and for 'insn1 r2'
would be '0010'.  Meanwhile 'insn2' would be '0011'.  Clearly we can
distinguish between these two instructions, however, right now, the cgen
disassembler can't.

If we are decoding '0011' but check insn1 first, then the opcode bits,
the '00__' match fine (as expected) then we extract '11' as the value
for the register field, after this we print the instruction as:

   insn1 ???

With this commit I introduce an operand validate phase.  After
extraction we now validate the operands.  For immediate / address
operands all values are considered valid, however, for registers, I
check that the value extracted is a valid register.  If it's not then I
indicate that the instruction was not a match.

This will be combined with changes in the opcode library, such that when
the extraction fails we continue looking for another instruction.

This is still a work in progress.  I think there are a couple of things
that should be improved.  These are:

   1. I think validation should be opt-in.  So operands that want
   validation should have to mark that they do in some way.

   2. It would be nice if the code to validate the operand could be
   overridden, imagine if _some_ immediate values were not used, but
   instead indicated a completely different instruction.

   3. If we find a matching instruction, but fail the validation, then
   go on searching, but don't find _any_ better match, it might be nice
   if the disassembler still printed the instruction with the '???'.  To
   do this we might need to have some kind of best-match-so-far
   mechanism, which we use for printing if we get no exact match.  Maybe
   even so far as having the validation "score" matches to allow the
   closest match to be selected.

ChangeLog:

	* opc-ibld.scm (/gen-validate-switch): New function.
	(/gen-extract-switch): Add call to C function created by
	/gen-validate-switch.
	(cgen-ibld.in) Call /gen-validate-switch.
	* opcodes.scm: Add gen-validate methods for <ifield>,
	<multi-ifield>, <derived-ifield>, <hw-index>, <hw-asm>, <keyword>,
	<hw-register>, <hw-immediate>, <hw-address>, <operand>, and
	<derived-operand>