Preserve an older version of the aburgess-scratch branch
The history of the aburgess-scratch branch has gotten lost in the
mists of time....

I found this version of the branch in another repository, and it's not
obvious if the changes between this branch and aburgess-scratch are
good or bad.

I suspect that the aburgess-scratch branch is better, and that this is
an older version.